+++ src/Lifecycle/NullLifecycle.php
@@ -14,7 +14,6 @@
 namespace PMG\Queue\Lifecycle;
 
 use PMG\Queue\Consumer;
-use PMG\Queue\Message;
 use PMG\Queue\MessageLifecycle;
 
 /**
@@ -30,7 +29,7 @@ class NullLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function starting(Message $message, Consumer $consumer)
+    public function starting(object $message, Consumer $consumer)
     {
         // noop
     }
@@ -38,7 +37,7 @@ class NullLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function completed(Message $message, Consumer $consumer)
+    public function completed(object $message, Consumer $consumer)
     {
         // noop
     }
@@ -46,7 +45,7 @@ class NullLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function retrying(Message $message, Consumer $consumer)
+    public function retrying(object $message, Consumer $consumer)
     {
         // noop
     }
@@ -54,7 +53,7 @@ class NullLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function failed(Message $message, Consumer $consumer)
+    public function failed(object $message, Consumer $consumer)
     {
         // noop
     }
@@ -62,7 +61,7 @@ class NullLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function succeeded(Message $message, Consumer $consumer)
+    public function succeeded(object $message, Consumer $consumer)
     {
         // noop
     }