+++ src/MessageTrait.php
@@ -13,7 +13,10 @@
 
 namespace PMG\Queue;
 
-@trigger_error(sprintf('The "%s" trait is deprecated since pmg/queue 5.0, a message\'s FQCN as the message name is the default in 5.0', MessageTrait::class));
+@trigger_error(sprintf(
+    'The "%s" trait is deprecated since pmg/queue 5.0, a message\'s FQCN as the message name is the default in 5.0',
+    MessageTrait::class
+), E_USER_DEPRECATED);
 
 /**
  * ABC for messages, implements `getName` as the class name.
@@ -23,7 +26,7 @@ namespace PMG\Queue;
 trait MessageTrait
 {
     /**
-     * @see NamedMessage::getName
+     * @see Message::getName
      */
     public function getName() : string
     {