+++ src/Router/FallbackRouter.php
@@ -13,7 +13,6 @@
 
 namespace PMG\Queue\Router;
 
-use PMG\Queue\Message;
 use PMG\Queue\Router;
 
 /**
@@ -43,7 +42,7 @@ final class FallbackRouter implements \PMG\Queue\Router
     /**
      * {@inheritdoc}
      */
-    public function queueFor(Message $message) : ?string
+    public function queueFor(object $message) : ?string
     {
         return $this->wrapped->queueFor($message) ?: $this->fallbackQueue;
     }